requirements-macos.txt
pyautogui>=0.9.54
webcolors>=1.13
Pillow>=10.0.0
numpy>=1.24.0
mss>=7.0.1
